requirements.txt
# Core
Django>=4.2,<5.0
djangorestframework>=3.15,<4.0

# Database
psycopg2-binary>=2.9

# Filtering
django-filter>=23.0

# Environment management
python-decouple>=3.8

# Image handling (profile pictures, logos, banners)
Pillow>=10.0

# Async task queue (for sending invitation emails)
celery>=5.3
redis>=5.0          # Celery broker

# Dev / testing
pytest>=7.0
pytest-django>=4.7
factory-boy>=3.3    # test fixtures
coverage>=7.0
